# shop-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        description: 0=普通用户,1=管理员
        type: integer
      mobile:
        type: string
      nickname:
        type: string
      status:
        description: 0=禁用,1=正常
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 爱影视频APP商城系统和第三方站点管理API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 爱影视频商城API
  version: "1.0"
paths:
...
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取令牌
      parameters:
      - description: 登录信息
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 用户登录
      tags:
      - 用户
  /api/v1/users/password:
    put:
      consumes:
      - application/json
      description: 修改当前用户的密码
      parameters:
      - description: 密码信息
        in: body
        name: password
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 用户
  /api/v1/users/profile:
    get:
      consumes:
      - application/json
      description: 获取当前用户的资料
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户资料
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 更新当前用户的资料
      parameters:
      - description: 用户资料
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户资料
      tags:
      - 用户
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 注册信息
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: 用户注册
      tags:
      - 用户
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"